types(templates): separate templates into own folder
