+++ src/listeners/client/eventsub/streamOnline.ts
@@ -1,5 +1,5 @@
 import { HOST_TWITCH_ID } from "../../../config.js";
-import { createEventSubListener } from "../../../types/EventSubTemplate.js";
+import { createEventSubListener } from "../../../templates/EventSubTemplate.js";
 import { twitchLog } from "../../../utilities/logger.js";
 
 const eventSubStreamOnline = createEventSubListener({
