+++ src/listeners/interaction/commands/discord.ts
@@ -1,5 +1,6 @@
 import { DISCORD_INVITE } from "../../../config/constants.js";
-import { createCommand, PermissionFlags } from "../../../types/CommandTemplate.js";
+import { createCommand, PermissionFlags } from "../../../templates/CommandTemplate.js";
+import { LinkFormatType } from "../../../types/LinkFormatType.js";
 
 const discordCommand = createCommand({
   name: "discord",
